+++ lib/src/log_options.dart
@@ -9,6 +9,13 @@ typedef MessageFormatting = Object Function(
   DateTime dateTime,
 );
 
+/// Override output message formatting
+typedef OverrideLoggerOutput = String? Function(
+  Object message,
+  LogLevel logLevel,
+  DateTime dateTime,
+);
+
 /// Logger options
 @immutable
 abstract base class LogOptions {
@@ -18,6 +25,7 @@ abstract base class LogOptions {
     bool printColors,
     bool outputInRelease,
     MessageFormatting? messageFormatting,
+    OverrideLoggerOutput? overrideOutput,
   }) = _LogOptionsImpl;
 
   const LogOptions._();
@@ -34,8 +42,15 @@ abstract base class LogOptions {
   /// Output message formatting callback
   MessageFormatting? get messageFormatting;
 
+  /// Output message formatting callback, allow to implement custom formatting
+  /// and your own printing logic for output messages.
+  /// If returns `null` then the message
+  /// will not be output by the default logger.
+  /// E.g. useful for JSON output, or output to a file/database/server.
+  OverrideLoggerOutput? get overrideOutput;
+
   /// Default Logger options
-  static const defaultOptions = _LogOptionsImpl();
+  static const LogOptions defaultOptions = _LogOptionsImpl();
 }
 
 final class _LogOptionsImpl extends LogOptions {
@@ -44,6 +59,7 @@ final class _LogOptionsImpl extends LogOptions {
     this.printColors = true,
     this.outputInRelease = false,
     this.messageFormatting,
+    this.overrideOutput,
   }) : super._();
 
   @override
@@ -57,4 +73,7 @@ final class _LogOptionsImpl extends LogOptions {
 
   @override
   final MessageFormatting? messageFormatting;
+
+  @override
+  final OverrideLoggerOutput? overrideOutput;
 }